Moved MaterialIconButton into its own file; Implemented basic download progress
